Hollander–McKeague confidence bands

Computes the Hollander–McKeague likelihood-ratio simultaneous confidence band¹ for the Kaplan–Meier survival estimator at each unique event time.

Usage

WH_HollanderMcKeague(
  time,
  risk,
  event,
  alpha = 0.05,
  verbose = TRUE,
  adapt = FALSE,
  adapt_alpha = 0.001,
  tol_G = 1e-10,
  maxit_G = 10000L,
  tol_K = 1e-10,
  maxit_K = 10000L
)

Arguments

time

Numeric vector of unique event times.

risk

Integer vector of numbers at risk at each time.

event

Integer vector of numbers of events at each time.

alpha

Optional significance level. Default = 0.05.

verbose

Optional logical; if TRUE, prints additional diagnostic information. Default = TRUE.

adapt

Optional logical; if TRUE, iteratively excludes the latest remaining event time from the time range of the confidence band until a desired pointwise αp exceeding adapt_alpha is obtained. Default = FALSE.

adapt_alpha

Optional minimum acceptable value for the pointwise αp if adapt = TRUE. Default = 0.001.

tol_G

Optional convergence tolerance for the infinite series computing the Hall–Wellner² distribution function G(a, λ). Default = 1e-10.

maxit_G

Optional maximum number of iterations for the infinite series computing the Hall–Wellner distribution function G(a, λ). Default = 10000.

tol_K

Optional convergence tolerance for the bisection root-finding of the Hall–Wellner quantile function K(x, a). Default = 1e-10.

maxit_K

Optional maximum number of iterations for the bisection root-finding of the Hall–Wellner quantile function K(x, a). Default = 10000.

Value

A numeric matrix with 2 columns: lower limit and upper limit.

Details

The Hollander–McKeague procedure begins with a specified simultaneous α and computes a smaller pointwise αp such that the resulting pointwise confidence intervals achieve the desired simultaneous coverage. When the tail of the survival curve is close to zero, αp can become very small and result in an undesirably wide confidence band. The adapt option addresses this by iteratively excluding the latest remaining event time from the time range of the confidence band until a desired αp exceeding adapt_alpha is obtained. Extremely small αp can cause 1 − αp to round to 1 in double precision; to prevent this, αp is always rounded up to half the machine epsilon.

References

Hollander, M. and McKeague, I.W., 1997. Likelihood ratio-based confidence bands for survival functions. Journal of the American Statistical Association, 92(437), pp. 215–226.

Hall, W.J. and Wellner, J.A., 1980. Confidence bands for a survival curve from censored data. Biometrika, 67(1), pp. 133–143.