Hollander–McKeague pointwise significance level

Computes the pointwise αp for Hollander–McKeague confidence bands.

Usage

WH_HollanderMcKeague_alpha(
  risk,
  event,
  surv,
  SE,
  alpha = 0.05,
  tol_G = 1e-10,
  maxit_G = 10000L,
  tol_K = 1e-10,
  maxit_K = 10000L
)

Arguments

risk

Integer vector of numbers at risk at each unique event time.

event

Integer vector of numbers of events at each unique event time.

surv

Numeric vector of Kaplan–Meier survival estimates at each unique event time.

SE

Numeric vector of standard errors for each surv.

alpha

Optional significance level. Default = 0.05.

tol_G

Optional convergence tolerance for the infinite series computing the Hall–Wellner² distribution function G(a, λ). Default = 1e-10.

maxit_G

Optional maximum number of iterations for the infinite series computing the Hall–Wellner distribution function G(a, λ). Default = 10000.

tol_K

Optional convergence tolerance for the bisection root-finding of the Hall–Wellner quantile function K(x, a). Default = 1e-10.

maxit_K

Optional maximum number of iterations for the bisection root-finding of the Hall–Wellner quantile function K(x, a). Default = 10000.

Value

The pointwise significance level αp.

References

Hollander, M. and McKeague, I.W., 1997. Likelihood ratio-based confidence bands for survival functions. Journal of the American Statistical Association, 92(437), pp. 215–226.

Hall, W.J. and Wellner, J.A., 1980. Confidence bands for a survival curve from censored data. Biometrika, 67(1), pp. 133–143.