Nair confidence bands

Computes the Nair's log-transformed equal-precision simultaneous confidence band¹ for the Kaplan–Meier survival estimator at each unique event time.

Usage

WH_Nair(
  time,
  surv,
  SE,
  risk,
  event,
  alpha = 0.05,
  verbose = TRUE,
  tol = 1e-10,
  maxit = 10000L,
  adapt = TRUE,
  MC_step = 100000L,
  MC_rep = 100000L,
  MC_seed = 24601L,
  e_alpha_override = 0
)

Arguments

time

Numeric vector of unique event times.

surv

Numeric vector of Kaplan–Meier survival estimates at each time.

SE

Numeric vector of standard errors for each surv.

risk

Integer vector of numbers at risk at each time.

event

Integer vector of numbers of events at each time.

alpha

Optional significance level. Default = 0.05.

verbose

Optional logical; if TRUE, prints additional diagnostic information. Default = TRUE.

tol

Optional convergence tolerance for the Borokov–Sycheva approximation. Default = 1e-10.

maxit

Optional maximum number of iterations for the Borokov–Sycheva approximation. Default = 10000.

adapt

Optional logical; if TRUE, computation of eα automatically falls back to the Monte Carlo method if the Borokov–Sycheva approximation fails to converge. Default = TRUE.

MC_step

Optional number of variance-stabilized Brownian bridge discretization steps for the Monte Carlo fallback. Default = 1e5.

MC_rep

Optional number of replicates for the Monte Carlo fallback. Default = 1e5.

MC_seed

Optional random seed for the Monte Carlo fallback. Default = 24601.

e_alpha_override

Optional numeric value overriding the critical value eα.

Value

A numeric matrix with 3 columns: time, lower limit, and upper limit.

Details

The equal-precision critical value eα is calculated using Borokov–Sycheva approximation² (WH_e_alpha()). This method aligns more closely with Monte Carlo simulations using 100,000 variance-stabilized Brownian bridge discretization steps and 100,000 replicates (WH_e_alpha_MC()) compared to the pre-computed tables published in Klein and Moeschberger (2003)³.

References

Nair, V.N., 1984. Conﬁdence bands for survival functions with censored data: a comparative study. Technometrics, 26, pp. 265–275.

Borokov, A.A. and Sycheva, N.M., 1968. On asymptotically optimal non-parametric criteria. Theory of Probability & Its Applications, 13(3), pp. 359–393.

Klein, J.P. and Moeschberger, M.L., 2003. Appendix C: Statistical Tables. In Survival Analysis: Techniques for Censored and Truncated Data, pp. 455–482. New York: Springer New York.